+++ index.html
@@ -250,8 +250,7 @@
     </section>
   </template>
 
-  <script src="js/main.js" defer></script>
-  <script src="js/functions.js" defer></script>
+  <script src="js/main.mjs" type="module"></script>
 </body>
 
 </html>
