+++ js/id-generator.mjs
@@ -1,0 +1,12 @@
+// Функция генерирует id
+
+const createIdGenerator = () => {
+  let lastGeneratedId = 0;
+  return () => {
+    lastGeneratedId += 1;
+    return lastGeneratedId;
+  };
+};
+
+export { createIdGenerator };
+
