+++ js/main.mjs
@@ -1,0 +1,7 @@
+import { createphotoDescriptions } from './data.mjs';
+import { validateInputLength, isPalindrome, extractNumbers } from './input-validation.mjs';
+
+const photoDescriptions = createphotoDescriptions();
+
+// Проверка
+// console.log(photoDescriptions);
